split logic in LongLatItem class into 3 separate classes
